Refactored test_utils.py into 3 files - we can better test new coming comp models
